// src/types/XykServiceTypes.d.ts
export interface UniswapLikeEcosystemCharts {
    /** * The name of the DEX, eg: `uniswap_v2`. */
    dex_name: string;
    /** * The requested chain ID eg: `1`. */
    chain_id: string;
    /** * The requested quote currency eg: `USD`. */
    quote_currency: string;
    gas_token_price_quote: number;
    total_swaps_24h: number;
    total_active_pairs_7d: number;
    total_fees_24h: number;
    /** * A prettier version of the gas quote for rendering purposes. */
    pretty_gas_token_price_quote: string;
    /** * A prettier version of the 24h total fees for rendering purposes. */
    pretty_total_fees_24h: string;
    volume_chart_7d: VolumeEcosystemChart[];
    volume_chart_30d: VolumeEcosystemChart[];
    liquidity_chart_7d: LiquidityEcosystemChart[];
    liquidity_chart_30d: LiquidityEcosystemChart[];
}
export interface VolumeEcosystemChart {
    /** * The name of the DEX, eg: `uniswap_v2`. */
    dex_name: string;
    /** * The requested chain ID eg: `1`. */
    chain_id: string;
    dt: Date;
    /** * The requested quote currency eg: `USD`. */
    quote_currency: string;
    volume_quote: number;
    /** * A prettier version of the volume quote for rendering purposes. */
    pretty_volume_quote: string;
    swap_count_24: number;
}
export interface LiquidityEcosystemChart {
    /** * The name of the DEX, eg: `uniswap_v2`. */
    dex_name: string;
    /** * The requested chain ID eg: `1`. */
    chain_id: string;
    dt: Date;
    /** * The requested quote currency eg: `USD`. */
    quote_currency: string;
    liquidity_quote: number;
    /** * A prettier version of the liquidity quote for rendering purposes. */
    pretty_liquidity_quote: string;
}